approve_matches: Support different print formats
Added option --format which allows to print the positive answers in
several formats: "clips" and "clips_comments" which are about to be
fed to video-composer and "subtitles" which is supposed to be a help
when manually editing the clips.

Added option --no-sort which allows to not sort the printed answers.
Default is now to sort the answers case-insensitive (was no sorting).